+++ src/diagnosis_plots.py
@@ -23,6 +23,7 @@ sns.set()
 import umap
 from distinctipy import distinctipy
 from scipy.interpolate import griddata
+from sklearn.decomposition import PCA
 
 
 
@@ -177,7 +178,7 @@ def rawInputUMAP(spatial_df, scRNA_df, scRNA_celltype, plot_colors):
     # relplot return a FacetGrid object
     # specify figure size by Height (in inches) of each facet, and Aspect ratio of each facet
     fgrid = sns.relplot(data=plot_df, x='UMAP1', y='UMAP2', hue='celltype', size='datatype', style='datatype', sizes=plot_sizes, markers=plot_markers, palette=plot_colors, kind='scatter', col='dataset', col_order=['scRNA-seq', 'spatial'], height=4.8*2, aspect=6.4/4.8)
-    fgrid.set(xlabel='Embedding Dimension 1', ylabel='Embedding Dimension 2')
+    fgrid.set(xlabel='UMAP 1', ylabel='UMAP 2')
     # Put the legend out of the figure
     #ax.legend(loc='center left', bbox_to_anchor=(1, 0.5))
     
@@ -357,7 +358,7 @@ def diagnosisCVAE(cvae, encoder, decoder, spatial_embed, spatial_transformed_df,
     # relplot return a FacetGrid object
     # specify figure size by Height (in inches) of each facet, and Aspect ratio of each facet
     fgrid = sns.relplot(data=plot_df, x='UMAP1', y='UMAP2', hue='celltype', s=20, marker='o', palette=plot_colors, kind='scatter', col='dataset', col_order=['scRNA-seq', 'spatial'], height=4.8*2, aspect=6.4/4.8)
-    fgrid.set(xlabel='Embedding Dimension 1', ylabel='Embedding Dimension 2')
+    fgrid.set(xlabel='UMAP 1', ylabel='UMAP 2')
     # Put the legend out of the figure
     #ax.legend(loc='center left', bbox_to_anchor=(1, 0.5))
     
@@ -365,11 +366,78 @@ def diagnosisCVAE(cvae, encoder, decoder, spatial_embed, spatial_transformed_df,
     plt.close()
     
     
+    # save UMAP coordinates of latent space
+    plot_df.loc[plot_df['datatype']=='spatial-spot', ['UMAP1', 'UMAP2']].to_csv(os.path.join(diagnosis_path, 'CVAE_latent_space', 'UMAP_coordinates_latent_mu_embedding_spatial_spots.csv.gz'), compression='gzip')
+    # save scRNA-seq cells UMAP coordinates
+    plot_df.loc[plot_df['datatype']=='scrna-cell', ['UMAP1', 'UMAP2']].to_csv(os.path.join(diagnosis_path, 'CVAE_latent_space', 'UMAP_coordinates_latent_mu_embedding_scRNA-seq_cells.csv.gz'), compression='gzip')
+    
+    
+    # plot PCA of latent space of spatial spots and scRNA-seq cells plus pseudo spots
+    # the order will affect the point overlay, first row draw first
+    all_pca = PCA(n_components=2).fit_transform(np.concatenate((pseudo_spot_embed, pseudo_spatial_embed, scRNA_embed, spatial_embed), axis=0))
+    
+    # add cell/spot count in the annotation
+    plot_df = pd.DataFrame({'PC1': all_pca[:, 0],
+                            'PC2': all_pca[:, 1],
+                            'celltype': ['pseudo']*pseudo_spot_embed.shape[0] +
+                                        ['pseudo']*pseudo_spatial_embed.shape[0] + 
+                                        [f'{x} ({celltype_count_dict[x]})' for x in scRNA_celltype.celltype.to_list()] +
+                                        [f'spatial ({spatial_embed.shape[0]})']*spatial_embed.shape[0],
+                            'dataset': ['scRNA-seq']*pseudo_spot_embed.shape[0] +
+                                       ['spatial']*pseudo_spatial_embed.shape[0] +
+                                       ['scRNA-seq']*scRNA_embed.shape[0] +
+                                       ['spatial']*spatial_embed.shape[0],
+                            'datatype': ['scrna-pseudo']*pseudo_spot_embed.shape[0] +
+                                        ['spatial-pseudo']*pseudo_spatial_embed.shape[0] +
+                                        ['scrna-cell']*scRNA_embed.shape[0] +
+                                        ['spatial-spot']*spatial_embed.shape[0],
+                            },
+                           index = [f'scrna_pseudo{x}' for x in range(pseudo_spot_embed.shape[0])] +
+                                   [f'spatial_pseudo{x}' for x in range(pseudo_spatial_embed.shape[0])] +
+                                   scRNA_decode_df.index.to_list() +
+                                   spatial_transformed_df.index.to_list())
+    
+    # plot colors already defined
+    sns.set_style("darkgrid")
+    
+    # relplot return a FacetGrid object
+    # specify figure size by Height (in inches) of each facet, and Aspect ratio of each facet
+    fgrid = sns.relplot(data=plot_df, x='PC1', y='PC2', hue='celltype', s=20, marker='o', palette=plot_colors, kind='scatter', col='dataset', col_order=['scRNA-seq', 'spatial'], height=4.8*2, aspect=6.4/4.8)
+    fgrid.set(xlabel='PC 1', ylabel='PC 2')
+    # Put the legend out of the figure
+    #ax.legend(loc='center left', bbox_to_anchor=(1, 0.5))
+    
+    plt.savefig(os.path.join(diagnosis_path, 'CVAE_latent_space', 'PCA_latent_mu_embedding_color_by_celltype.png'))
+    plt.close()
+    
+    
+    # save PCA coordinates of latent space
+    plot_df.loc[plot_df['datatype']=='spatial-spot', ['PC1', 'PC2']].to_csv(os.path.join(diagnosis_path, 'CVAE_latent_space', 'PCA_coordinates_latent_mu_embedding_spatial_spots.csv.gz'), compression='gzip')
+    # save scRNA-seq cells PCA coordinates
+    plot_df.loc[plot_df['datatype']=='scrna-cell', ['PC1', 'PC2']].to_csv(os.path.join(diagnosis_path, 'CVAE_latent_space', 'PCA_coordinates_latent_mu_embedding_scRNA-seq_cells.csv.gz'), compression='gzip')
+    
+    
+    # Plot PCA density
+    # for safe to avoid program exit when density estimation failed
+    try:
+        plt.figure(figsize=(6.4*2, 4.8*2))
+        sns.set_style('whitegrid')
+        ax = sns.kdeplot(data=plot_df, x='PC1', y='PC2', fill=True)
+        ax.set(xlabel='PC 1', ylabel='PC 2')
+        plt.savefig(os.path.join(diagnosis_path, 'CVAE_latent_space', 'PCA_latent_mu_embedding_color_by_density.png'))
+        plt.close()
+    except:
+        pass
+    
+    
+    # UPDATE: change to draw on PCA instead of UMAP
     # plot distribution of number of cells in pseudo-spots
     # add number of cells in pseudo-spot to dataframe
     plot_df['n_cell_in_spot'] = np.nan
     plot_df.loc[plot_df['datatype']=='scrna-cell', 'n_cell_in_spot'] = 1
     plot_df.loc[plot_df['datatype']=='scrna-pseudo', 'n_cell_in_spot'] = n_cell_in_spot
+    
+    sns.set_style("darkgrid")
  
     # generate a colormap with a specified color for NA (spatial spots), but not work for relplot...
     # show the full legend of colorbar in relplot, otherwise it will only show a sample of evenly spaced values (The FacetGrid hue is categorical, not continuous)
@@ -377,30 +445,31 @@ def diagnosisCVAE(cvae, encoder, decoder, spatial_embed, spatial_transformed_df,
     fig, (ax1, ax2) = plt.subplots(1, 2, sharey=True, sharex=True, figsize=(6.4*4, 4.8*2))
     # left panel: scatter plot of pseudo-spots + scRNA-seq cells
     tmp_df = plot_df.loc[plot_df['dataset']=='scRNA-seq']
-    sc = ax1.scatter(tmp_df['UMAP1'], tmp_df['UMAP2'], c=tmp_df['n_cell_in_spot'], cmap='cubehelix', s=10, marker='o')
+    sc = ax1.scatter(tmp_df['PC1'], tmp_df['PC2'], c=tmp_df['n_cell_in_spot'], cmap='cubehelix', s=10, marker='o')
     ax1.set_title('dataset = scRNA-seq')
-    ax1.set_xlabel('Embedding Dimension 1')
-    ax1.set_ylabel('Embedding Dimension 2')
+    ax1.set_xlabel('PC 1')
+    ax1.set_ylabel('PC 2')
     
     # right panel: scatter plot of spatial spots + spatial pseudo-spots
     # first draw spatial pseudo-spots, which lay at the bottom
     tmp_df = plot_df.loc[plot_df['datatype']=='spatial-pseudo']
-    ax2.scatter(tmp_df['UMAP1'], tmp_df['UMAP2'], color=plot_colors['pseudo'], s=10, marker='o')
+    ax2.scatter(tmp_df['PC1'], tmp_df['PC2'], color=plot_colors['pseudo'], s=10, marker='o')
     tmp_df = plot_df.loc[plot_df['datatype']=='spatial-spot']
-    ax2.scatter(tmp_df['UMAP1'], tmp_df['UMAP2'], color=plot_colors[f'spatial ({spatial_embed.shape[0]})'], s=10, marker='o')
+    ax2.scatter(tmp_df['PC1'], tmp_df['PC2'], color=plot_colors[f'spatial ({spatial_embed.shape[0]})'], s=10, marker='o')
     ax2.set_title('dataset = spatial')
-    ax2.set_xlabel('Embedding Dimension 1')
-    ax2.set_ylabel('Embedding Dimension 2')
+    ax2.set_xlabel('PC 1')
+    ax2.set_ylabel('PC 2')
     
     # add colorbar with title to the most right (https://stackoverflow.com/questions/13784201/how-to-have-one-colorbar-for-all-subplots, conflict with tight_layout)
     cbar = fig.colorbar(sc, ax=ax2)
     cbar.ax.set_title('#cell in spot')
     
     fig.tight_layout()
-    fig.savefig(os.path.join(diagnosis_path, 'CVAE_latent_space', 'UMAP_latent_mu_embedding_color_by_num_cell_in_spot.png'))
+    fig.savefig(os.path.join(diagnosis_path, 'CVAE_latent_space', 'PCA_latent_mu_embedding_color_by_num_cell_in_spot.png'))
     plt.close()
     
     
+    # UPDATE: change to draw on PCA instead of UMAP
     # plot distribution of cell-type proportions of each cell-type
     # create a proportion dataframe with the same row order
     prop_df = pd.concat([pseudo_spots_celltype_prop,
@@ -412,7 +481,7 @@ def diagnosisCVAE(cvae, encoder, decoder, spatial_embed, spatial_transformed_df,
     assert (prop_df.index == plot_df.index).all()
     
     # save multiple figures into one PDF
-    with PdfPages(os.path.join(diagnosis_path, 'CVAE_latent_space', 'UMAP_latent_mu_embedding_color_by_celltype_proportion.pdf')) as pdf:
+    with PdfPages(os.path.join(diagnosis_path, 'CVAE_latent_space', 'PCA_latent_mu_embedding_color_by_celltype_proportion.pdf')) as pdf:
         for this_celltype in celltype_order:
         
             # start to plot
@@ -423,16 +492,16 @@ def diagnosisCVAE(cvae, encoder, decoder, spatial_embed, spatial_transformed_df,
             # don't forget the .values
             tmp_df = tmp_df.assign(proportion = prop_df.loc[plot_df['dataset']=='scRNA-seq', this_celltype].values)
             
-            sc = ax1.scatter(tmp_df['UMAP1'], tmp_df['UMAP2'], c=tmp_df['proportion'], cmap='cubehelix', s=10, marker='o', norm=Normalize(vmin=0, vmax=1))
+            sc = ax1.scatter(tmp_df['PC1'], tmp_df['PC2'], c=tmp_df['proportion'], cmap='cubehelix', s=10, marker='o', norm=Normalize(vmin=0, vmax=1))
        
             ax1.set_title('dataset = scRNA-seq')
-            ax1.set_xlabel('Embedding Dimension 1')
-            ax1.set_ylabel('Embedding Dimension 2')
+            ax1.set_xlabel('PC 1')
+            ax1.set_ylabel('PC 2')
         
             # right panel: scatter plot of spatial spots + spatial pseudo-spots
             # we also interpolate the grid and draw a contour plot of cell type proportions in right panel
-            grid_x, grid_y = np.mgrid[tmp_df['UMAP1'].min():tmp_df['UMAP1'].max():0.025, tmp_df['UMAP2'].min():tmp_df['UMAP2'].max():0.025]
-            grid_z = griddata(tmp_df.loc[:, ['UMAP1', 'UMAP2']].values, tmp_df['proportion'].values, (grid_x, grid_y), method='linear',  fill_value=np.nan)
+            grid_x, grid_y = np.mgrid[tmp_df['PC1'].min():tmp_df['PC1'].max():0.025, tmp_df['PC2'].min():tmp_df['PC2'].max():0.025]
+            grid_z = griddata(tmp_df.loc[:, ['PC1', 'PC2']].values, tmp_df['proportion'].values, (grid_x, grid_y), method='linear',  fill_value=np.nan)
         
             try:
                 ax2.contourf(grid_x, grid_y, grid_z, cmap='cubehelix', norm=Normalize(vmin=0, vmax=1), alpha=0.3)
@@ -441,13 +510,13 @@ def diagnosisCVAE(cvae, encoder, decoder, spatial_embed, spatial_transformed_df,
         
             # first draw spatial pseudo-spots, which lay at the bottom
             tmp_df = plot_df.loc[plot_df['datatype']=='spatial-pseudo']
-            ax2.scatter(tmp_df['UMAP1'], tmp_df['UMAP2'], color=plot_colors['pseudo'], s=10, marker='o')
+            ax2.scatter(tmp_df['PC1'], tmp_df['PC2'], color=plot_colors['pseudo'], s=10, marker='o')
             tmp_df = plot_df.loc[plot_df['datatype']=='spatial-spot']
-            ax2.scatter(tmp_df['UMAP1'], tmp_df['UMAP2'], color=plot_colors[f'spatial ({spatial_embed.shape[0]})'], s=10, marker='o')
+            ax2.scatter(tmp_df['PC1'], tmp_df['PC2'], color=plot_colors[f'spatial ({spatial_embed.shape[0]})'], s=10, marker='o')
         
             ax2.set_title('dataset = spatial')
-            ax2.set_xlabel('Embedding Dimension 1')
-            ax2.set_ylabel('Embedding Dimension 2')
+            ax2.set_xlabel('PC 1')
+            ax2.set_ylabel('PC 2')
         
             # add colorbar with title
             cbar = fig.colorbar(sc, ax=ax2)
@@ -459,12 +528,6 @@ def diagnosisCVAE(cvae, encoder, decoder, spatial_embed, spatial_transformed_df,
             
             pdf.savefig(fig)  # Save the current figure to pdf
             plt.close(fig)  # Close the figure to free memory
-    
-    
-    # save UMAP coordinates of latent space
-    plot_df.loc[plot_df['datatype']=='spatial-spot', ['UMAP1', 'UMAP2']].to_csv(os.path.join(diagnosis_path, 'CVAE_latent_space', 'UMAP_coordinates_latent_mu_embedding_spatial_spots.csv.gz'), compression='gzip')
-    # save scRNA-seq cells UMAP coordinates
-    plot_df.loc[plot_df['datatype']=='scrna-cell', ['UMAP1', 'UMAP2']].to_csv(os.path.join(diagnosis_path, 'CVAE_latent_space', 'UMAP_coordinates_latent_mu_embedding_scRNA-seq_cells.csv.gz'), compression='gzip')
     
     
     # UMAP of decoded spatial and scRNA-seq cell gene expression
@@ -505,7 +568,7 @@ def diagnosisCVAE(cvae, encoder, decoder, spatial_embed, spatial_transformed_df,
     # relplot return a FacetGrid object
     # specify figure size by Height (in inches) of each facet, and Aspect ratio of each facet
     fgrid = sns.relplot(data=plot_df, x='UMAP1', y='UMAP2', hue='celltype', size='datatype', style='datatype', sizes=plot_sizes, markers=plot_markers, palette=plot_colors, kind='scatter', col='dataset', col_order=['scRNA-seq', 'spatial'], height=4.8*2, aspect=6.4/4.8)
-    fgrid.set(xlabel='Embedding Dimension 1', ylabel='Embedding Dimension 2')
+    fgrid.set(xlabel='UMAP 1', ylabel='UMAP 2')
     # Put the legend out of the figure
     #ax.legend(loc='center left', bbox_to_anchor=(1, 0.5))